requirements.txt
Flask==1.0.2
mysql-connector==2.2.9
Jinja2==2.10
